+++ apps/www/src/lib/bundlephobia.ts
@@ -8,16 +8,14 @@ export function formatBytes(bytes: number, decimals = 2) {
   const sizes = ['Bytes', 'kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
   const i = Math.floor(Math.log(bytes) / Math.log(k));
 
-  return (
-    Number.parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i]
-  );
+  return `${Number.parseFloat((bytes / k ** i).toFixed(dm))} ${sizes[i]}`;
 }
 
 export async function getPackageData(name: string) {
   // export async function getPackageData(name, version) {
   const bundlephobiaResponse = await fetch(
-    // `https://bundlephobia.com/api/size?package=@udecode/${name}@${version}`
-    `https://bundlephobia.com/api/size?package=@udecode/${name}`
+    // `https://bundlephobia.com/api/size?package=@platejs/${name}@${version}`
+    `https://bundlephobia.com/api/size?package=@platejs/${name}`
   );
 
   // sometimes we get an empty response body