+++ packages/core/src/static/__tests__/render.spec.tsx
@@ -1,27 +1,22 @@
 import React from 'react';
 import { serializeHtml } from '../serializeHtml';
 import { createStaticEditor } from './create-static-editor';
-import { BaseEditorKit } from '@/registry/components/editor/editor-base-kit';
 import { createTSlatePlugin, createSlateEditor } from '../../lib';
+import { BaseEditorKit } from '../../../../../apps/www/src/registry/components/editor/editor-base-kit';
 
 describe('serializePlateStatic nodes', () => {
   it('should serialize render below nodes', async () => {
     const renderBelowPlugin = createTSlatePlugin({
       key: 'test-list',
       render: {
-        belowNodes: (injectProps: any) => {
-          return function Component({
-            children,
-          }: {
-            children: React.ReactNode;
-          }) {
+        belowNodes: (_injectProps: any) =>
+          function Component({ children }: { children: React.ReactNode }) {
             return (
               <ul>
                 <li>{children}</li>
               </ul>
             );
-          };
-        },
+          },
       },
     });