+++ packages/link/src/lib/withLink.spec.tsx
@@ -103,13 +103,17 @@ describe('withLink', () => {
         ) as any;
 
         it('should insert link', () => {
-          jest.spyOn(JSON, 'parse').mockReturnValue(<fragment>docs</fragment>);
+          const jsonParseSpy = spyOn(JSON, 'parse').mockReturnValue(
+            <fragment>docs</fragment>
+          );
 
           const editor = createTestEditor(input);
 
           editor.tf.insertData(data);
 
           expect(editor.children).toEqual(output.children);
+
+          jsonParseSpy.mockRestore();
         });
       });
 
@@ -180,13 +184,17 @@ describe('withLink', () => {
             </editor>
           ) as any;
 
-          jest.spyOn(JSON, 'parse').mockReturnValue(<fragment>docs</fragment>);
+          const jsonParseSpy = spyOn(JSON, 'parse').mockReturnValue(
+            <fragment>docs</fragment>
+          );
 
           const editor = createTestEditor(input);
 
           editor.tf.insertData(data);
 
           expect(editor.children).toEqual(output.children);
+
+          jsonParseSpy.mockRestore();
         });
       });
     });
@@ -420,9 +428,7 @@ describe('withLink', () => {
             plugins: [
               BaseLinkPlugin.configure({
                 options: {
-                  getUrlHref: (_url) => {
-                    return 'http://google.com';
-                  },
+                  getUrlHref: (_url) => 'http://google.com',
                 },
               }),
             ],