+++ packages/udecode/react-hotkeys/src/__tests__/HotkeysProvider.test.tsx
@@ -1,11 +1,11 @@
+/// <reference types="@testing-library/jest-dom" />
+
 import React from 'react';
-import type { ReactNode } from 'react';
 
 import { act, render, renderHook } from '@testing-library/react';
-
-import type { HotkeysContextType } from '../internal/HotkeysProvider';
-
+import type { ReactNode } from 'react';
 import { HotkeysProvider, useHotkeys, useHotkeysContext } from '..';
+import type { HotkeysContextType } from '../internal/HotkeysProvider';
 
 it('should render children', () => {
   const { getByText } = render(
@@ -14,7 +14,7 @@ it('should render children', () => {
     </HotkeysProvider>
   );
 
-  expect(getByText('Hello')).toBeInTheDocument();
+  expect(getByText('Hello')).toBeTruthy();
 });
 
 it('should default to wildcard scope', () => {
@@ -163,13 +163,11 @@ it('should update bound hotkeys when useHotkeys changes its scopes', () => {
     return useHotkeysContext();
   };
 
-  const wrapper = ({ children }: { children: ReactNode }) => {
-    return (
-      <HotkeysProvider initiallyActiveScopes={['foo']}>
-        {children}
-      </HotkeysProvider>
-    );
-  };
+  const wrapper = ({ children }: { children: ReactNode }) => (
+    <HotkeysProvider initiallyActiveScopes={['foo']}>
+      {children}
+    </HotkeysProvider>
+  );
 
   const { rerender, result } = renderHook<
     HotkeysContextType,
@@ -178,7 +176,6 @@ it('should update bound hotkeys when useHotkeys changes its scopes', () => {
     initialProps: {
       scopes: ['foo'],
     },
-    // @ts-ignore
     wrapper,
   });