+++ packages/basic-nodes/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @platejs/basic-nodes
 
+## 52.0.1
+
+### Patch Changes
+
+- [#4750](https://github.com/udecode/plate/pull/4750) by [@zbeyens](https://github.com/zbeyens) – Add React Compiler support.
+
+## 52.0.0
+
+### Major Changes
+
+- [#4747](https://github.com/udecode/plate/pull/4747) by [@zbeyens](https://github.com/zbeyens) – ESM-only
+
+## 51.1.2
+
+### Patch Changes
+
+- [#4732](https://github.com/udecode/plate/pull/4732) by [@zbeyens](https://github.com/zbeyens) – Format code with Biome
+
 ## 51.0.0
 
 ## 49.0.0
@@ -21,25 +39,25 @@
     - Configure shortcuts manually via the `shortcuts` field in plugin configuration.
     - Example (Block Plugins):
       ```ts
-      H1Plugin.configure({ shortcuts: { toggle: { keys: 'mod+alt+1' } } });
+      H1Plugin.configure({ shortcuts: { toggle: { keys: "mod+alt+1" } } });
       BlockquotePlugin.configure({
-        shortcuts: { toggle: { keys: 'mod+shift+period' } },
+        shortcuts: { toggle: { keys: "mod+shift+period" } },
       });
       ```
     - Example (Mark Plugins):
       ```ts
-      CodePlugin.configure({ shortcuts: { toggle: { keys: 'mod+e' } } });
+      CodePlugin.configure({ shortcuts: { toggle: { keys: "mod+e" } } });
       StrikethroughPlugin.configure({
-        shortcuts: { toggle: { keys: 'mod+shift+x' } },
+        shortcuts: { toggle: { keys: "mod+shift+x" } },
       });
       SubscriptPlugin.configure({
-        shortcuts: { toggle: { keys: 'mod+comma' } },
+        shortcuts: { toggle: { keys: "mod+comma" } },
       });
       SuperscriptPlugin.configure({
-        shortcuts: { toggle: { keys: 'mod+period' } },
+        shortcuts: { toggle: { keys: "mod+period" } },
       });
       HighlightPlugin.configure({
-        shortcuts: { toggle: { keys: 'mod+shift+h' } },
+        shortcuts: { toggle: { keys: "mod+shift+h" } },
       });
       ```