+++ packages/dnd/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @platejs/dnd
 
+## 52.0.1
+
+### Patch Changes
+
+- [#4750](https://github.com/udecode/plate/pull/4750) by [@zbeyens](https://github.com/zbeyens) – Add React Compiler support.
+
+## 52.0.0
+
+### Major Changes
+
+- [#4747](https://github.com/udecode/plate/pull/4747) by [@zbeyens](https://github.com/zbeyens) – ESM-only
+
+## 51.1.3
+
+### Patch Changes
+
+- [#4735](https://github.com/udecode/plate/pull/4735) by [@zbeyens](https://github.com/zbeyens) – Avoid accessing ref during render
+
+## 51.1.2
+
+### Patch Changes
+
+- [#4732](https://github.com/udecode/plate/pull/4732) by [@zbeyens](https://github.com/zbeyens) – Format code with Biome
+
 ## 51.0.0
 
 ## 50.3.1